Lift error level to 3 and treat warnings as errors

+++ ctt/Player.cpp
@@ -150,7 +150,7 @@ void Player::addScrubber(VideoScrubber::sptr scrubber, unsigned int position) {
 	{
 		throw new IllegalArgumentException("Tried to add a dummy VideoScrubber to a player.");
 	}
-	if (position > videoScrubbers.size())
+	if (position > static_cast<unsigned int>(videoScrubbers.size()))
 	{
 		throw new IllegalArgumentException("Tried to insert a Scrubber into a Player at position " 
 			+ QString::number(position) + ", but " + QString::number(videoScrubbers.size()) 
@@ -168,7 +168,7 @@ void Player::removeScrubber(unsigned int position) {
 	{
 		throw new IllegalStateException("Tried to add a VideoScrubber to a dummy player.");
 	}
-	if (position >= videoScrubbers.size())
+    if (position >= static_cast<unsigned int>(videoScrubbers.size()))
 	{
 		throw new IllegalArgumentException("Tried to remove a Scrubber from a Player at position "
 			+ QString::number(position) + ", but it contained only " + QString::number(videoScrubbers.size())
